+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/jetpack/JetpackActivationFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.jetpackinstallation
+package com.woocommerce.android.ui.login.jetpack
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,7 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 
-class JetpackConfigurationFragment : Fragment() {
+class JetpackActivationFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         return ComposeView(requireContext()).apply {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
