+++ WooCommerce/src/main/res/layout/fragment_login_jetpack_activation_container.xml
@@ -6,4 +6,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:defaultNavHost="true"
-    app:navGraph="@navigation/nav_graph_jetpack_configuration" />
+    app:navGraph="@navigation/nav_graph_jetpack_activation" />
