--- WooCommerce/src/main/res/navigation/nav_graph_jetpack_configuration.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_graph_jetpack_configuration"
-    app:startDestination="@id/jetpackConfigurationFragment">
-
-    <fragment
-        android:id="@+id/jetpackConfigurationFragment"
-        android:name="com.woocommerce.android.ui.login.jetpackinstallation.JetpackConfigurationFragment"
-        android:label="JetpackConfigurationFragment" />
-</navigation>
